+++ src/molecules/Button/Button.stories.tsx
@@ -1,27 +1,27 @@
 import { ChangeEvent, useState } from 'react';
 
-import {
-  Checkbox,
-  EdsProvider,
-  Icon,
-  Snackbar,
-  Tooltip,
-} from '@equinor/eds-core-react';
-import { add, menu, refresh, save } from '@equinor/eds-icons';
-import { Meta, StoryFn, StoryObj } from '@storybook/react-vite';
+import { Checkbox, Snackbar, Tooltip } from '@equinor/eds-core-react';
+import { save } from '@equinor/eds-icons';
+import { Meta, StoryFn } from '@storybook/react-vite';
 
-import { Button, ButtonProps } from './Button';
-import page from 'src/molecules/Button/Button.docs.mdx';
+import { ButtonProps } from './Button';
+import { spacings } from 'src/atoms';
+import { Button } from 'src/molecules/Button/Button';
 import { Stack } from 'src/storybook';
-
-import { expect, fn, userEvent } from 'storybook/test';
 
 const meta: Meta<typeof Button> = {
   title: 'Molecules/Button',
   component: Button,
   args: {
-    as: 'button',
+    variant: 'filled',
+    color: 'primary',
     loading: undefined,
+  },
+  parameters: {
+    router: {
+      initialEntries: ['/'],
+      routes: ['$'],
+    },
   },
   argTypes: {
     loading: {
@@ -32,34 +32,72 @@ const meta: Meta<typeof Button> = {
     },
     disabled: { control: 'boolean', type: 'boolean' },
     variant: {
-      options: ['ghost', 'contained', 'outlined'],
+      options: ['filled', 'outlined', 'ghost'],
       control: {
         type: 'select',
       },
     },
-    as: {
-      options: ['span', 'a', 'button'],
+    color: {
+      options: ['primary', 'danger'],
       control: {
         type: 'select',
-      },
-    },
-  },
-  parameters: {
-    docs: {
-      page,
-      source: {
-        excludeDecorators: true,
       },
     },
   },
 };
 
 export default meta;
-
 export const Introduction: StoryFn<ButtonProps> = (args) => {
   return <Button {...args}>You can control me</Button>;
 };
 Introduction.decorators = [
+  (Story) => (
+    <Stack>
+      <Story />
+    </Stack>
+  ),
+];
+
+export const Basic: StoryFn<ButtonProps> = () => (
+  <div
+    style={{ display: 'flex', flexDirection: 'column', gap: spacings.medium }}
+  >
+    <div style={{ display: 'flex', gap: spacings.medium }}>
+      <Button variant="filled">Filled</Button>
+      <Button variant="outlined">Outlined</Button>
+      <Button variant="ghost">Ghost</Button>
+    </div>
+    <div style={{ display: 'flex', gap: spacings.medium }}>
+      <Button color="danger">Filled</Button>
+      <Button color="danger" variant="outlined">
+        Outlined
+      </Button>
+      <Button color="danger" variant="ghost">
+        Ghost
+      </Button>
+    </div>
+  </div>
+);
+Basic.decorators = [
+  (Story) => (
+    <Stack>
+      <Story />
+    </Stack>
+  ),
+];
+
+export const Disabled: StoryFn<ButtonProps> = () => (
+  <div style={{ display: 'flex', gap: spacings.medium }}>
+    <Button disabled>Filled</Button>
+    <Button disabled variant="outlined">
+      Outlined
+    </Button>
+    <Button disabled variant="ghost">
+      Ghost
+    </Button>
+  </div>
+);
+Disabled.decorators = [
   (Story) => (
     <Stack>
       <Story />
@@ -107,20 +145,15 @@ Accessibility.decorators = [
   ),
 ];
 
-export const Basic: StoryFn<ButtonProps> = () => (
+export const Icons: StoryFn<ButtonProps> = () => (
   <>
-    <Button>Contained</Button>
-    <Button variant="contained_icon" aria-label="add action">
-      <Icon data={add}></Icon>
-    </Button>
-    <Button variant="outlined">Outlined</Button>
-    <Button variant="ghost">Ghost</Button>
-    <Button variant="ghost_icon" aria-label="save action">
-      <Icon data={save}></Icon>
-    </Button>
+    <div style={{ display: 'flex', gap: spacings.medium }}>
+      <Button leadingIcon={save}>Leading icon</Button>
+      <Button trailingIcon={save}>Trailing icon</Button>
+    </div>
   </>
 );
-Basic.decorators = [
+Icons.decorators = [
   (Story) => (
     <Stack>
       <Story />
@@ -128,192 +161,49 @@ Basic.decorators = [
   ),
 ];
 
-export const LoadingState: StoryFn<ButtonProps> = () => (
-  <>
-    <Button loading>Contained</Button>
-    <Button variant="contained_icon" aria-label="add action" loading>
-      <Icon data={add}></Icon>
-    </Button>
-    <Button variant="outlined" loading>
-      Outlined
-    </Button>
-    <Button variant="ghost" loading>
-      Ghost
-    </Button>
-    <Button variant="ghost_icon" aria-label="save action" loading>
-      <Icon data={save}></Icon>
-    </Button>
-    <Button loading color="danger">
-      Contained
-    </Button>
-    <Button
-      variant="contained_icon"
-      color="danger"
-      aria-label="add action"
-      loading
-    >
-      <Icon data={add}></Icon>
-    </Button>
-    <Button variant="outlined" color="danger" loading>
-      Outlined
-    </Button>
-    <Button variant="ghost" color="danger" loading>
-      Ghost
-    </Button>
-    <Button
-      variant="ghost_icon"
-      aria-label="save action"
-      color="danger"
-      loading
-    >
-      <Icon data={save}></Icon>
-    </Button>
-  </>
+export const Loading: StoryFn<ButtonProps> = () => (
+  <div
+    style={{ display: 'flex', flexDirection: 'column', gap: spacings.medium }}
+  >
+    <div style={{ display: 'flex', gap: spacings.medium }}>
+      <Button loading>Filled</Button>
+      <Button loading variant="outlined">
+        Outlined
+      </Button>
+      <Button loading variant="ghost">
+        Ghost
+      </Button>
+    </div>
+    <div style={{ display: 'flex', gap: spacings.medium }}>
+      <Button loading color="danger">
+        Filled
+      </Button>
+      <Button loading color="danger" variant="outlined">
+        Outlined
+      </Button>
+      <Button loading color="danger" variant="ghost">
+        Ghost
+      </Button>
+    </div>
+  </div>
 );
-LoadingState.decorators = [
+Loading.decorators = [
   (Story) => (
     <Stack>
       <Story />
     </Stack>
   ),
 ];
-
-export const IconButton: StoryFn<ButtonProps> = () => (
-  <>
-    <Button variant="ghost_icon" aria-label="save action">
-      <Icon data={save}></Icon>
-    </Button>
-    <Button variant="ghost_icon" color="secondary" aria-label="save action">
-      <Icon data={save}></Icon>
-    </Button>
-    <Button variant="ghost_icon" color="danger" aria-label="save action">
-      <Icon data={save}></Icon>
-    </Button>
-    <Button variant="ghost_icon" disabled aria-label="save action">
-      <Icon data={save}></Icon>
-    </Button>
-    <Button variant="contained_icon" aria-label="add action">
-      <Icon data={add}></Icon>
-    </Button>
-    <Button variant="contained_icon" color="secondary" aria-label="add action">
-      <Icon data={add}></Icon>
-    </Button>
-    <Button variant="contained_icon" color="danger" aria-label="add action">
-      <Icon data={add}></Icon>
-    </Button>
-    <Button variant="contained_icon" disabled aria-label="add action">
-      <Icon data={add}></Icon>
-    </Button>
-  </>
-);
-
-IconButton.storyName = 'Icon button';
-IconButton.decorators = [
-  (Story) => (
-    <Stack>
-      <Story />
-    </Stack>
-  ),
-];
-
-export const Color: StoryFn<ButtonProps> = () => (
-  <>
-    <Button color="primary">Primary</Button>
-    <Button color="secondary">Secondary</Button>
-    <Button color="danger">Danger</Button>
-  </>
-);
-Color.decorators = [
-  (Story) => (
-    <Stack>
-      <Story />
-    </Stack>
-  ),
-];
-
-export const Hierarchy: StoryFn<ButtonProps> = () => (
-  <>
-    <Button>Contained</Button>
-    <Button variant="outlined">Outlined</Button>
-    <Button variant="ghost">Ghost</Button>
-  </>
-);
-Hierarchy.decorators = [
-  (Story) => (
-    <Stack>
-      <Story />
-    </Stack>
-  ),
-];
-
-export const Compact: StoryFn<ButtonProps> = () => {
-  const [compact, setComfortable] = useState<boolean>(true);
-
-  return (
-    <EdsProvider density={compact ? 'compact' : 'comfortable'}>
-      <Checkbox
-        label="Compact"
-        onChange={() => {
-          setComfortable(!compact);
-        }}
-        checked={compact}
-      />
-      <Stack direction="row">
-        <Button>Contained</Button>
-        <Button variant="outlined">Outlined</Button>
-        <Button variant="ghost">Ghost</Button>
-        <Button variant="ghost_icon" aria-label="menu action">
-          <Icon data={menu} title="Ghost icon menu"></Icon>
-        </Button>
-      </Stack>
-    </EdsProvider>
-  );
-};
-Compact.decorators = [
-  (Story) => (
-    <Stack direction="column">
-      <Story />
-    </Stack>
-  ),
-];
-
-export const WithTooltip: StoryFn<ButtonProps> = () => (
-  <>
-    <Tooltip title="This is what a tooltip looks like">
-      <Button>Hover me</Button>
-    </Tooltip>
-    <Tooltip title="This tooltip only shows for people using firefox and using mouse. Don't do this!">
-      <Button disabled>Disabled button</Button>
-    </Tooltip>
-    <Tooltip title="Tooltip works in all browsers and with keyboard navigation when using aria-disabled">
-      <Button aria-disabled>Aria-disabled button</Button>
-    </Tooltip>
-  </>
-);
-
-WithTooltip.decorators = [
-  (Story) => (
-    <Stack>
-      <Story />
-    </Stack>
-  ),
-];
-WithTooltip.storyName = 'Disabled buttons and tooltip';
 
 export const FullWidth: StoryFn<ButtonProps> = () => (
   <>
-    <Button fullWidth>
-      <Icon data={refresh} size={16}></Icon>
-      Fullwidth
+    <Button fullWidth leadingIcon={save}>
+      fullWidth
     </Button>
-    <Button fullWidth>
-      Fullwidth
-      <Icon data={refresh} size={16}></Icon>
+    <Button fullWidth trailingIcon={save}>
+      fullWidth
     </Button>
-    <Button>
-      <Icon data={refresh} size={16}></Icon>
-      not Fullwidth
-    </Button>
+    <Button leadingIcon={save}>No fullWidth</Button>
   </>
 );
 FullWidth.storyName = 'Full width';
@@ -324,98 +214,3 @@ FullWidth.decorators = [
     </div>
   ),
 ];
-
-export const All: StoryFn<ButtonProps> = () => (
-  <>
-    <Button>Primary</Button>
-    <Button color="secondary">Secondary</Button>
-    <Button color="danger">Danger</Button>
-    <Button disabled>Disabled</Button>
-    <Button variant="outlined">Primary</Button>
-    <Button variant="outlined" color="secondary">
-      Secondary
-    </Button>
-    <Button variant="outlined" color="danger">
-      Danger
-    </Button>
-    <Button variant="outlined" disabled>
-      Disabled
-    </Button>
-    <Button variant="ghost">Primary</Button>
-    <Button variant="ghost" color="secondary">
-      Secondary
-    </Button>
-    <Button variant="ghost" color="danger">
-      Danger
-    </Button>
-    <Button variant="ghost" disabled>
-      Disabled
-    </Button>
-    <Button variant="ghost_icon" aria-label="save action">
-      <Icon data={save} title="save action"></Icon>
-    </Button>
-    <Button variant="ghost_icon" color="secondary" aria-label="save action">
-      <Icon data={save}></Icon>
-    </Button>
-    <Button variant="ghost_icon" color="danger" aria-label="save action">
-      <Icon data={save}></Icon>
-    </Button>
-    <Button variant="ghost_icon" disabled aria-label="save action">
-      <Icon data={save}></Icon>
-    </Button>
-    <Button variant="contained_icon" aria-label="add action">
-      <Icon data={add}></Icon>
-    </Button>
-    <Button variant="contained_icon" color="secondary" aria-label="add action">
-      <Icon data={add}></Icon>
-    </Button>
-    <Button variant="contained_icon" color="danger" aria-label="add action">
-      <Icon data={add}></Icon>
-    </Button>
-    <Button variant="contained_icon" disabled aria-label="add action">
-      <Icon data={add}></Icon>
-    </Button>
-  </>
-);
-All.decorators = [
-  (Story) => (
-    <Stack
-      style={{
-        display: 'grid',
-        gridTemplateColumns: 'repeat(4, auto)',
-      }}
-    >
-      <Story />
-    </Stack>
-  ),
-];
-
-type Story = StoryObj<typeof Button>;
-
-export const TestLoadingState: Story = {
-  tags: ['test-only'],
-  args: {
-    loading: true,
-    children: 'Loading Button',
-    onClick: fn(),
-  },
-  play: async ({ canvas, args }) => {
-    await expect(canvas.getByRole('progressbar')).toBeInTheDocument();
-    await expect(canvas.queryByText('Loading Button')).not.toBeVisible();
-
-    await userEvent.click(canvas.getByRole('button'));
-    await expect(args.onClick).not.toHaveBeenCalled();
-  },
-};
-
-export const TestClickable: Story = {
-  tags: ['test-only'],
-  args: {
-    children: 'Click me',
-    onClick: fn(),
-  },
-  play: async ({ canvas, args }) => {
-    await userEvent.click(canvas.getByRole('button'));
-    await expect(args.onClick).toHaveBeenCalledOnce();
-  },
-};
