+++ src/molecules/Button/Button.styles.ts
@@ -1,3 +1,4 @@
+import { Icon } from '@equinor/eds-core-react';
 import { tokens } from '@equinor/eds-tokens';
 import { typographyTemplate } from '@equinor/eds-utils';
 
@@ -6,7 +7,7 @@ import { shape, spacings } from 'src/atoms';
 
 import { css, styled } from 'styled-components';
 
-interface ButtonStyles {
+export interface ButtonStyles {
   $tokens: VariantTokens;
   $fullWidth?: boolean;
 }
@@ -16,7 +17,7 @@ const resolveBorderColor = (tokens: {
   backgroundColor: string;
 }) => ('borderColor' in tokens ? tokens.borderColor : tokens.backgroundColor);
 
-export const BaseButton = styled.button<ButtonStyles>`
+export const ButtonPrimitive = styled.button<ButtonStyles>`
   display: inline-flex;
   align-items: center;
   justify-content: center;
@@ -93,4 +94,27 @@ export const BaseButton = styled.button<ButtonStyles>`
       }
     }
   }
+`;
+
+export const HiddenText = styled.span`
+  opacity: 0;
+  visibility: hidden;
+`;
+
+export const CenteredContent = styled.span`
+  left: 0;
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
+
+export const LeftIcon = styled(Icon)`
+  left: 0;
+`;
+
+export const RightIcon = styled(Icon)`
+  right: 0;
 `;
