+++ src/molecules/index.ts
@@ -7,6 +7,8 @@ export { Banner } from './Banner/Banner';
 export type { BannerProps } from './Banner/Banner';
 export { Button } from './Button/Button';
 export type { ButtonProps } from './Button/Button';
+export { IconButton } from './Button/IconButton/IconButton';
+export type { IconButtonProps } from './Button/IconButton/IconButton';
 export { Chip } from './Chip/Chip';
 export type { BaseChipProps } from './Chip/Chip';
 export { Confetti } from './Confetti/Confetti';
